start styling process
