+++ src/lib/NavButton.svelte
@@ -2,6 +2,6 @@
 	export let classes;
 </script>
 
-<button class="bg-green-400 rounded-sm p-2 h-10 {classes}">
+<button class="bg-green-400 rounded-sm p-2 shadow-2xl {classes}">
 	<slot />
 </button>
