+++ src/routes/_data/vivald copy.md
@@ -1,0 +1,14 @@
+---
+slug: vivald
+title: Vivaldi
+name: VIVALDI BROWSER
+published: true
+---
+
+For me, it's the <b name="insert">{name}</b>. The best internet browser. I even download the <b name="insert">{name}</b> on my phone, tablet, laptop, and desktop because the <b name="insert">{name}</b> is so user friendly and helps fill my workflow with awesome tab features (Tab Stacks & Tab Tilling)
+    
+One time I asked for 3 tab stacks, WITH the tab tilling (multiple sites open at once on a screen). I said, "Wow, three for free!" and the nice friendly <b name="insert">{name}</b> dev laughed and said, "I'm going to call you 3-for-free!". But the best part is with tab tilling, it's not even free for me!
+    
+Now the <b name="insert">{name}</b> greets me with "hey it's 3-for-free!" and ALWAYS gets me a tab stack, with a web panel on the side. It's such a fun and cool atmosphere with my <b name="insert">{name}</b>, I use it at least 7 times a week for my daily browsing needs which include work, shitposting, school, and stalking my ex on facebook (with her instagram tiled to the side!).
+    
+What a great <b name="insert">{name}</b>
