+++ src/routes/index.svelte
@@ -1,5 +1,5 @@
 <script context="module">
-	const pastaFiles = import.meta.glob("./pastas/*.md");
+	const pastaFiles = import.meta.glob("./_data/*.md");
 	let body = [];
 	for (const path in pastaFiles) {
 		body.push(pastaFiles[path]().then(({ metadata }) => metadata));
@@ -28,20 +28,23 @@
 		return comparison * -1;
 	}
 	pastas.sort(sortByDate);
+
+	import NavButton from "$lib/NavButton.svelte";
 </script>
 
 <div class="text-center mt-[15%] text-white">
 	<h1 class="font-normal text-3xl">The <span class="text-green-400 font-bold">BEST</span> copypasta generator for shilling browsers</h1>
-	<p>Choose a browser from the cards below to start <span>spammin'</span>.</p>
-</div>
+	<p class="leading-6 text-xl mt-1">Choose a browser from the cards below to start <span class="text-green-400 font-bold">spammin'</span>.</p>
 
-
-<div class="blog">
-	<h1>Posts</h1>
-	{#each pastas as { slug, title, summary, published }}
+	{#each pastas as { slug, title, published }}
 		{#if published}
-			<a rel="prefetch" href="{slug}" class="item-name">{title}</a>
-			<p class="item-desc">- {summary}</p>
+			<div class="ml-3 first:ml-0 inline-flex items-center mt-4 m-auto">
+				<a rel="prefetch" href="copypastas/{slug}">
+					<NavButton classes="text-xl font-medium bg-gray-600">
+						{title}
+					</NavButton>
+				</a>
+			</div>
 		{/if}
 	{/each}
 </div>
