--- src/routes/pastas/h.md
@@ -1,16 +1,0 @@
----
-slug: h
-title: Vivaldi
-summary: null
-date: null
-author: null
-published: true
----
-
-For me, it's the <b name="insert">BRAVE BROWSER</b>. The best internet browser. I even download the <b name="insert">BRAVE BROWSER</b> on my phone, tablet, laptop, and desktop because the <b name="insert">BRAVE BROWSER</b> is so user friendly and helps fill my wallets with BAT (Basic Attention Token)
-
-One time I asked for 3 ads to be served and they gave me three, WITH the BAT (Basic Attention Token). I said, "Wow, three for free!" and the nice friendly <b name="insert">BRAVE BROWSER</b> dev laughed and said, "I'm going to call you 3-for-free!". But the best part is with BAT, it's not even free for me!
-
-Now the <b name="insert">BRAVE BROWSER</b> greets me with "hey it's 3-for-free!" and ALWAYS blocks every ad, cookie, and tracker. It's such a fun and cool atmosphere with my <b name="insert">BRAVE BROWSER</b>,  I use it at least 7 times a week for my daily browsing needs which include work, shitposting, school, and stalking my ex on facebook (without the trackers!).
-
-What a great <b name="insert">BRAVE BROWSER</b>
